* Rotation is fixed, pieces rotate around where you hold them. * Stopwatch bug fixed, now displays every second

+++ src/main/java/Game.java
@@ -74,7 +74,7 @@ public class Game {
         stopwatch.resume();
     }
 
-    public long getElapsedSeconds() {
+    public int getElapsedSeconds() {
         return stopwatch.getElapsedTime();
     }
     public boolean savePlayerBoard() throws IOException {
